requirements.txt
streamlit
PIL
tensorflow
keras
efficientnet
numpy
pandas
matplotlib
plotly
opencv-python
tqdm
imutils
scikit-learn
Pillow
ipykernel
google-generativeai
pymongo
python-dotenv
torch
torchvision